fix: rename the folder
