+++ user-app/src/app/page.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image"
-import Dashboard from "./dashboard/page"
+import Dashboard from "./home/page"
 
 export default function Home() {
   return <Dashboard />
